Refactor abstract LocationDataSource
