+++ app/src/main/java/com/godamy/myplayer/model/LocationDataSource.kt
@@ -1,0 +1,7 @@
+package com.godamy.myplayer.model
+
+import android.location.Location
+
+interface LocationDataSource {
+    suspend fun findLastLocation(): Location?
+}
