+++ app/src/main/java/com/godamy/myplayer/model/RegionRepository.kt
@@ -1,18 +1,14 @@
 package com.godamy.myplayer.model
 
 import android.Manifest
-import android.annotation.SuppressLint
 import android.location.Geocoder
 import android.location.Location
 import androidx.appcompat.app.AppCompatActivity
-import com.google.android.gms.location.LocationServices
-import kotlinx.coroutines.suspendCancellableCoroutine
-import kotlin.coroutines.resume
 
 class RegionRepository(activity: AppCompatActivity) {
 
-    // Inicializo el location service de google con el activity
-    private val fusedLocationClient = LocationServices.getFusedLocationProviderClient(activity)
+    // TODO IOC
+    private val locationDataSource: LocationDataSource = PlayServiceLocationDataSource(activity)
     private val geocoder = Geocoder(activity)
 
     private val coarsePermisionChecker =
@@ -23,16 +19,8 @@ class RegionRepository(activity: AppCompatActivity) {
     // Set default location when is false
     private suspend fun findLastLocation(): Location? {
         val success = coarsePermisionChecker.request()
-        return if (success) getRegionFromLocation() else null
+        return if (success) locationDataSource.findLastLocation() else null
     }
-
-    @SuppressLint("MissingPermission")
-    private suspend fun getRegionFromLocation(): Location? =
-        suspendCancellableCoroutine { continuation ->
-            fusedLocationClient.lastLocation.addOnCompleteListener {
-                continuation.resume(it.result)
-            }
-        }
 
     private fun Location?.toRegion(): String {
         val address = this?.let {
